feat(authentication): 新增refresh_token redis存储和有效性校验

+++ centaur-authentication/authentication-infrastructure/src/main/java/com/sky/centaur/authentication/infrastructure/token/gatewayimpl/redis/TokenRepository.java
@@ -25,6 +25,6 @@ import com.sky.centaur.authentication.infrastructure.token.gatewayimpl.redis.dat
  * @since 1.0.0
  */
 public interface TokenRepository extends
-    RedisDocumentRepository<TokenRedisDo, Integer> {
+    RedisDocumentRepository<TokenRedisDo, Long> {
 
 }
